clean up generated code and tests
The Generated code was using MooseX::AttributeHelpers which has been
folded into core. We fix that to use Native::Attribute's now. We also
move to use nammespace::clean.

We also start moving the tests to use `done_testing` instead of no_plan.

Finally we fix the Filter to properly add the XML trait to text nodes as
well as non-text ones, this squishes some warnings.