* add test for yp-api
We needed to have RDF support for the YP Feed API. This adds a test that slurps the current version of the API XML so we can round-trip it.

* get the XML tests passing

The `namespace_map` for the Generator has to be different from the ones for the Loader and Builder. We fix the test for this, fix the code to reflect the changes, and everything magically works.

* clean up the logic from the unresolve_namespaces flow

change the logic (and wording) to "unless no_unresolved_namespaces" rather than "if unresovled_namespaces" since we're checking if the list "is_empty"