fix to run with the current Moose and MooseX::MetaDescription
Moose 0.94 introduced some changes to the way Attributes are handled in
Roles. As a side effect the Attribute inheritance no longer works in
Role -> Role composition. We copy the Attribute now until this is fixed
in Moose.

Also patch a few tests to deal with the changes that have happened in
the codebase that we didn't detect before.