build a new example based on XSPF
going through and building a new example using the XSPF file for a
client, I had to perform some cleanup of how generate and the
templates work.

The templates have been unified into XML::Toolkit;:Cmd::ClassTemplate
with the generate command wrapping it with the logic to break stuff out.
This cleans up some nastiness and makes the templates more uniform.

we're also adding a test for XSPF round tripping.